Add start script to package.json for Express app
This commit adds a "start" script to the package.json file. This allows developers to easily start the Express application by running 'npm start' in the terminal.

By doing so, we standardize the way the application is started, making it easier for new developers to get up to running with the codebase.

The start script simply runs 'node index.js', which is the main entry point of our Express application.